o Migrated to Eclipse Aether

+++ maven-core/src/test/java/org/apache/maven/repository/TestRepositoryConnectorFactory.java
@@ -20,11 +20,11 @@ package org.apache.maven.repository;
  */
 
 import org.codehaus.plexus.component.annotations.Component;
-import org.sonatype.aether.RepositorySystemSession;
-import org.sonatype.aether.repository.RemoteRepository;
-import org.sonatype.aether.spi.connector.RepositoryConnector;
-import org.sonatype.aether.spi.connector.RepositoryConnectorFactory;
-import org.sonatype.aether.transfer.NoRepositoryConnectorException;
+import org.eclipse.aether.RepositorySystemSession;
+import org.eclipse.aether.repository.RemoteRepository;
+import org.eclipse.aether.spi.connector.RepositoryConnector;
+import org.eclipse.aether.spi.connector.RepositoryConnectorFactory;
+import org.eclipse.aether.transfer.NoRepositoryConnectorException;
 
 /**
  * @author Benjamin Bentmann
@@ -40,7 +40,7 @@ public class TestRepositoryConnectorFactory
         return new TestRepositoryConnector( repository );
     }
 
-    public int getPriority()
+    public float getPriority()
     {
         return 0;
     }
